# Web/Controller/UserController.cs
using Domain.Dtos.UserDto;
using Domain.Filter;
using Infrastructure.Interface;
using Infrastructure.Response;
using Infrastructure.Seed;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;

namespace Web.Controller;
[ApiController]
[Route("api/[controller]")]
public class UserController(IUserService service) : ControllerBase
{
    [HttpGet]
    [Authorize(Roles = Roles.Admin)]
    public async Task<PaginationResponse<List<GetUserDto>>> GetAll([FromQuery] UserFilter filter) =>
        await service.GetAllUserAsync(filter);
    
    [HttpPut("{id}")]
    [Authorize(Roles = Roles.Admin)]
    public async Task<ApiResponse<string>> Update([FromRoute] string id, [FromBody] UpdateUserDto request) =>
        await service.UpdateAsync(id, request);
                                                                                                          
    [HttpDelete("{id}")]
    [Authorize(Roles = Roles.Admin)]
    public async Task<ApiResponse<string>> Delete([FromRoute] string id) => await service.DeleteAsync(id);
}